Python 3/requirements.txt
pycrypto==2.6.1
gmpy2==2.0.8
pwntools==4.0.1
tqdm==4.45.0